# chunker speedup