I'm a bit late to the https://github.com/borgbackup/borg/issues/1021 party, buy here's a chunker speedup by rearranging the loop a bit. Testing throughput on my slow old laptop:

Before:

```(borg-env) lenovo_0$ borg-env/bin/python3  ~/borgbench
    length   count      dt throughput chunks_count
1000000000       1 5.442 183.741    120
 100000000      10 5.531 180.795     12
  10000000     100 5.225 191.399      2
   1000000    1000 2.764 361.778      1
    100000   10000 0.172 5826.037      1
     10000  100000 0.993 1006.679      1
      1000 1000000 8.999 111.122      1
       100 1000000 8.776 11.394      1
        10 1000000 8.763 1.141      1
         1 1000000 8.768 0.114      1
```

After:
```(borg-env) lenovo_0$ borg-env/bin/python3  ~/borgbench
    length   count      dt throughput chunks_count
1000000000       1 3.891 256.990    120
 100000000      10 3.989 250.687     12
  10000000     100 3.736 267.687      2
   1000000    1000 2.022 494.470      1
    100000   10000 0.172 5827.721      1
     10000  100000 0.984 1016.437      1
      1000 1000000 8.931 111.970      1
       100 1000000 8.682 11.518      1
        10 1000000 8.649 1.156      1
         1 1000000 8.677 0.115      1
```