@ThomasWaldmann 2019-11-09
BTW, an important test after finishing the optimization work would be to make sure that the new code cuts precisely the same chunks as the old code. This is important for deduplication in existing repos.