@ncleaton 2019-11-10
> BTW, an important test after finishing the optimization work would be to make sure that the new code cuts precisely the same chunks as the old code. This is important for deduplication in existing repos.

Good point. Should a script to perform this test go into the borg repo, either as part of the test suite (perhaps normally disabled) or as something in the scripts directory ?