@ThomasWaldmann 2019-11-10
Yes, a unit test would be fine.