@ncleaton 2019-11-11
> Yes, a unit test would be fine.

Added. I've also run the same test code over each of the files under /boot on my system with old and new chunker code, and the results are identical.