@ncleaton 2019-11-12
Good feedback, thanks.

I'm not sure if I should amend the commits on my branch and force push, or just add a new commit with the feedback fixes. Do you have a preference ?