@ThomasWaldmann 2019-11-12
Additional commits are easier to review on github.