@ThomasWaldmann 2019-11-18
travis is unhappy, the blake2b import does not work.