@ThomasWaldmann 2019-11-21
there is a "self test count" mismatch now, this is why most tests now fail.

can you search for this phrase and adjust it to 38 (+1), please?