@ncleaton 2019-11-23
Fixed now, the 2 travis failures on the latest run were both out of time in archiver.py before it got to my new test.