@ThomasWaldmann 2019-11-23
Restarted the 2 macOS jobs, somehow that stuff was extremely slow (but usually is only slow).