@ThomasWaldmann 2019-11-26
Hmm, I recently noticed that we run into timeout issues with the macOS based tests on travis-ci.

That is totally unrelated to this PR as it is not merged yet.

But while thinking about it, I noticed that if that new chunker test is included in the self-tests (which are executed at **each** borg start), it will make the tests even slower (and not just 1 test time slower, but N test times slower).

So, I guess while we want to have that test in our testsuite, we do not want to have it in the self-test tests.