@ThomasWaldmann 2019-11-28
Looks like the travis-ci macOS tests have a performance issue (in general) as they time out all the time currently.

Looks like this is ready to merge (I guess I'ld squash all commits together into one)?