@ThomasWaldmann 2019-11-29
Yay, we have a faster chunker now! Merged, thanks for your work!